Implement the rewrite-free dependant AST
The older rewrite system based versions of the compiler were ludicrously
slow when it came to compiling larger expressions.
By using a dependant AST where the expression type and its scope is
encoded into the AST type, we're able to fold expressions directly into LLVM.
This is also safer than any method I've previously attempted, not only
with regards to expression construction, but also when folding because
we no longer need to use error values in unexpected situations; we can
just prove that such situations are absurd.

The only downside to this new approach is that it requires an unusually
large quantity of proofs to convince the compiler that transformations
are correct.

We also need a bit of unsafe code to avoid a Peano representation of
natural numbers at runtime and gain even better performance.