Share bodies instead of copying them
This fixes a major problem with the previous version: there exist
pathological programs that lead to an exponential amount of copying.
This improvement comes with the small cost that non-pathological
programs take longer to compile.

Sharing is observed by first converting the program into an interaction
net, which is then reduced into the output program.
The reduction algorithm used is that of the Lambdascope implementation
extended with FFI nodes and the necessary rules to reduce them.
https://web.archive.org/web/20170706084403/http://www.phil.uu.nl/~oostrom/publication/pdf/lambdascope.pdf

For ease of debugging, a generic backend language was also introduced to
decouple most of the compiler from LLVM. This may also, in future, allow
alternative backends to be developed.